# api_server_fastapi/src/openapi_server/models/geometry.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.geometry_coordinates_inner import GeometryCoordinatesInner


class Geometry(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Geometry - a model defined in OpenAPI

        type: The type of this Geometry [Optional].
        coordinates: The coordinates of this Geometry [Optional].
    """

    type: Optional[str] = Field(alias="Type", default=None)
    coordinates: Optional[List[GeometryCoordinatesInner]] = Field(alias="Coordinates", default=None)

Geometry.update_forward_refs()